+++ src/include/LSP/Utils.hpp
@@ -4,7 +4,9 @@
 #include <filesystem>
 #include <fstream>
 #include <sstream>
+#include <vector>
 #include <unordered_map>
+#include <algorithm>
 
 std::optional<std::string> getParentPath(const std::string& path);
 std::optional<std::string> getAncestorPath(const std::string& path, const std::string& ancestorName);
@@ -17,6 +19,12 @@ std::string& toLower(std::string& str);
 bool endsWith(const std::string_view& str, const std::string_view& suffix);
 bool replace(std::string& str, const std::string& from, const std::string& to);
 void replaceAll(std::string& str, const std::string& from, const std::string& to);
+
+template<typename V>
+inline bool contains(const std::vector<V>& vec, const V& value)
+{
+    return std::find(std::begin(vec), std::end(vec), value) != std::end(vec);
+}
 
 template<class K, class V>
 inline bool contains(const std::unordered_map<K, V>& map, const K& value)